Separate unit-tests and test-tube tests for cl-vault (#682)
