+++ smart-contracts/contracts/cl-vault/.cargo/config
@@ -2,5 +2,5 @@
 wasm = "build --release --lib --target wasm32-unknown-unknown"
 unit-test = "test --lib"
 schema = "run --bin schema"
-test-tube = "test --package cl-vault --lib -- --include-ignored test_tube:: --nocapture --test-threads=1"
+test-tube = "test --test * --features test-tube -- --include-ignored --test-threads=1"
 test-tube-build = "build --release --lib --target  wasm32-unknown-unknown --target-dir ./test-tube-build"
