+++ smart-contracts/contracts/cl-vault/Cargo.toml
@@ -16,10 +16,9 @@ crate-type = ["cdylib", "rlib"]
 name = "schema"
 
 [features]
-# for more explicit tests, cargo test --features=backtraces
 backtraces = ["cosmwasm-std/backtraces"]
-# use library feature to disable all instantiate/execute/query exports
 library = []
+test-tube = ["osmosis-test-tube"]
 
 [dependencies]
 cosmwasm-std = { workspace = true }
@@ -37,7 +36,7 @@ num_enum = { workspace = true }
 apollo-cw-asset = { workspace = true }
 dex-router-osmosis = {workspace = true}
 cw-vault-multi-standard = {git = "https://github.com/quasar-finance/cw-vault-standard", branch ="master", features = ["lockup", "force-unlock"]}
+osmosis-test-tube = { workspace = true, optional = true }
 
 [dev-dependencies]
-osmosis-test-tube = { workspace = true }
 proptest = { workspace = true }
