+++ smart-contracts/contracts/cl-vault/src/contract.rs
@@ -301,13 +301,12 @@ pub fn migrate(deps: DepsMut, _env: Env, msg: MigrateMsg) -> Result<Response, Co
 mod tests {
     use super::*;
     #[allow(deprecated)]
+    use crate::state::USER_REWARDS;
+    #[allow(deprecated)]
     use crate::{
         helpers::coinlist::CoinList,
         state::{OldPosition, OldVaultConfig, Position, OLD_POSITION, POSITION},
-        test_tube::initialize::initialize::{DENOM_BASE, DENOM_QUOTE, DENOM_REWARD},
     };
-    #[allow(deprecated)]
-    use crate::{state::USER_REWARDS, test_tube::initialize::initialize::MAX_SLIPPAGE_HIGH};
     use cosmwasm_std::{
         coin,
         testing::{mock_dependencies, mock_env},
@@ -316,6 +315,11 @@ mod tests {
     use osmosis_std::{cosmwasm_to_proto_coins, types::cosmos::bank::v1beta1::MsgMultiSend};
     use prost::Message;
     use std::str::FromStr;
+
+    const DENOM_BASE: &str = "uatom";
+    const DENOM_QUOTE: &str = "ubtc";
+    const DENOM_REWARD: &str = "ustrd";
+    const MAX_SLIPPAGE_HIGH: u64 = 9000;
 
     pub fn mock_migrate(
         deps: DepsMut,
