+++ smart-contracts/contracts/cl-vault/src/lib.rs
@@ -2,7 +2,7 @@ pub mod contract;
 mod error;
 pub mod helpers;
 mod instantiate;
-mod math;
+pub mod math;
 pub mod msg;
 pub mod query;
 mod reply;
@@ -10,9 +10,6 @@ pub mod state;
 mod vault;
 
 pub use crate::error::ContractError;
-
-#[cfg(test)]
-mod test_tube;
 
 #[cfg(test)]
 mod test_helpers;
