--- smart-contracts/contracts/cl-vault/src/test_tube/mod.rs
@@ -1,10 +1,0 @@
-mod admin;
-mod any_deposit;
-mod authz;
-mod autocompound;
-mod deposit_withdraw;
-pub mod helpers;
-pub mod initialize;
-mod proptest;
-mod range;
-mod rewards;
